+++ backend/ianalyzer/settings_test.py
@@ -9,7 +9,7 @@ def test_corpus_path(*path):
 
 CORPORA = {
     'small-mock-corpus': test_corpus_path('small', 'small_mock_corpus.py'),
-    'large-mock-corpus': path_in_testdir('visualization', 'mock_corpora', 'large_mock_corpus.py'),
+    'large-mock-corpus': test_corpus_path('large', 'large_mock_corpus.py'),
     'multilingual-mock-corpus': path_in_testdir('download', 'mock_corpora', 'multilingual_mock_corpus.py'),
     'times': os.path.join(BASE_DIR, 'corpora', 'times', 'times.py'),
     'media-mock-corpus': path_in_testdir('media', 'media_mock_corpus.py'),
